Add docs to line rail, Move render code to helper, Changed dots to have edge colors
Before we were doing edge colors by adding more dots. This took extra work and could easily result in bad data. As each point was a new data point for rendering. Now we just define a color and edge color as needed. This improves the code and allow all dots to have edges making it easier to see.